+++ data_provider/__init__.py
@@ -42,6 +42,7 @@ from .longbridge_fetcher import LongbridgeFetcher
 from .finnhub_fetcher import FinnhubFetcher
 from .alphavantage_fetcher import AlphaVantageFetcher
 from .tickflow_fetcher import TickFlowFetcher
+from .kpl_fetcher import KplFetcher
 from .us_index_mapping import is_us_index_code, is_us_stock_code, get_us_index_yf_symbol, US_INDEX_MAPPING
 
 __all__ = [
@@ -58,6 +59,7 @@ __all__ = [
     'FinnhubFetcher',
     'AlphaVantageFetcher',
     'TickFlowFetcher',
+    'KplFetcher',
     'is_us_index_code',
     'is_us_stock_code',
     'is_hk_stock_code',